Save File and Save File As
Changed the "Save File" functionality to subscribe automatically the file that was opened. At the same time, it was added a "Save File as" function that does exactly what the previous function did. Implemented a dialog box asking if user wants to 'Save File' before opening another file. Only asks this if it's not the first file opened.